dbUpdate/requirements.txt
pandas~=1.2.4
pytest~=6.2.4
psycopg2~=2.8.6
python-dotenv~=0.17.1